Adds env command
