+++ SoftLayer/shell/cmd_env.py
@@ -1,0 +1,14 @@
+"""Print help text."""
+# :license: MIT, see LICENSE for more details.
+
+import click
+
+from SoftLayer.CLI import environment
+from SoftLayer.CLI import formatting
+
+
+@click.command()
+@environment.pass_env
+def cli(env):
+    """Print shell help text."""
+    return formatting.iter_to_table(env.vars)
